fix(core): CRLF normalization, empty-MIME inference, 64-bit sizeBytes

+++ src/main/kotlin/com/appfeedback/core/ParsedModels.kt
@@ -4,7 +4,7 @@ data class ParsedAttachment(
     val filename: String,
     val mimeType: String,
     val url: String,
-    val sizeBytes: Int?,
+    val sizeBytes: Long?,
 )
 
 data class ParsedFeedbackBody(
